requirements.txt
# PatientSim Requirements
# Updated for new model APIs

# LLM API clients
openai>=1.12.0
requests>=2.31.0

# Data handling
pyyaml>=6.0
jsonlines>=4.0.0

# Progress bars
tqdm>=4.66.0

# Analysis (optional, for existing analysis.ipynb)
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
